# backend/tests/openapi_snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Strom API",
    "description": "REST and WebSocket API for the Strom pipeline engine",
    "contact": {
      "name": "Per Enstedt",
      "email": "per.enstedt@eyevinn.se"
    },
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.3.26"
  },
  "paths": {
...
    "/whip/{endpoint_id}": {
      "post": {
        "tags": [
          "whip"
        ],
        "summary": "Handle WHIP POST request (SDP offer from client).",
        "description": "Proxies the SDP offer to the internal whipserversrc HTTP server and returns\nthe SDP answer, rewriting the Location header to use the proxy path.\n\nIf the internal server returns 500 or times out (stale session), we recreate\nthe whipserversrc element in the background and return the error immediately.\nThe client's own retry logic will resend the offer to the fresh element.",
        "operationId": "whip_post",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHIP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "WHIP session created, SDP answer returned",
            "content": {
              "application/sdp": {}
            }
          },
          "404": {
            "description": "WHIP endpoint not found"
          },
          "502": {
            "description": "Proxy error forwarding to internal WHIP server"
          },
          "503": {
            "description": "WHIP element busy, retry in a moment"
          }
        }
      },
      "options": {
        "tags": [
          "whip"
        ],
        "summary": "Handle CORS preflight for WHIP endpoints.",
        "operationId": "whip_options",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHIP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "CORS preflight response"
          }
        }
      }
    },
    "/whip/{endpoint_id}/resource/{resource_id}": {
      "delete": {
        "tags": [
          "whip"
        ],
        "summary": "Handle WHIP DELETE request (client disconnect).",
        "description": "After forwarding DELETE to whipserversrc, we trigger async element\nrecreation. The whipserversrc element is treated as single-use: it handles\none session, then gets destroyed and replaced with a fresh element.\nThe pad-removed callback also triggers recreation, but the AtomicBool\nidempotency flag in WhipServerContext prevents double-recreation.",
        "operationId": "whip_resource_delete",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHIP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "WHIP resource/session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "WHIP session deleted and element recreated"
          },
          "404": {
            "description": "WHIP endpoint not found"
          },
          "502": {
            "description": "Proxy error forwarding to internal WHIP server"
          }
        }
      },
      "options": {
        "tags": [
          "whip"
        ],
        "summary": "Handle CORS preflight for WHIP resource endpoints.",
        "operationId": "whip_resource_options",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHIP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "WHIP resource/session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "CORS preflight response"
          }
        }
      },
      "patch": {
        "tags": [
          "whip"
        ],
        "summary": "Handle WHIP PATCH request (ICE trickle from client).",
        "operationId": "whip_resource_patch",
        "parameters": [
          {
            "name": "endpoint_id",
            "in": "path",
            "description": "WHIP endpoint identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource_id",
            "in": "path",
            "description": "WHIP resource/session identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "ICE candidates accepted"
          },
          "404": {
            "description": "WHIP endpoint not found"
          },
          "502": {
            "description": "Proxy error forwarding to internal WHIP server"
          }
        }
      }
    }
  },
...
}